requirements.txt
PyQt5
mp-api
python-dotenv
db-sqlite3
keyring